+++ lib/messenger.dart
@@ -1,0 +1,100 @@
+part of 'short_navigation.dart';
+
+abstract class GoMessenger {
+  ///If you wnat to [showAdaptiveDialog] without using context
+  ///you have to use this function and the package will pass the context automaticlly.
+  ///You can use all [showAdaptiveDialog] proparties as usuall.
+  static Future<T?> dialog<T>(
+    Widget content, {
+    bool barrierDismissible = true,
+    Color? barrierColor,
+    String? barrierLabel,
+    bool useSafeArea = true,
+    bool useRootNavigator = true,
+    RouteSettings? routeSettings,
+    Offset? anchorPoint,
+    TraversalEdgeBehavior? traversalEdgeBehavior,
+  }) async {
+    return showAdaptiveDialog<T>(
+      context: Go.context,
+      builder: (context) => content,
+      barrierDismissible: barrierDismissible,
+      barrierColor: barrierColor,
+      barrierLabel: barrierLabel,
+      useSafeArea: useSafeArea,
+      useRootNavigator: useRootNavigator,
+      routeSettings: routeSettings,
+      anchorPoint: anchorPoint,
+      traversalEdgeBehavior: traversalEdgeBehavior,
+    );
+  }
+
+  ///If you wnat to [showModalBottomSheet] without using context
+  ///you have to use this function and the package will pass the context automaticlly.
+  ///You can use all [showModalBottomSheet] proparties as usuall.
+  static Future<T?> bottomSheet<T>(
+    Widget content, {
+    Color? backgroundColor,
+    String? barrierLabel,
+    double? elevation,
+    ShapeBorder? shape,
+    Clip? clipBehavior,
+    BoxConstraints? constraints,
+    Color? barrierColor,
+    bool isScrollControlled = false,
+    double scrollControlDisabledMaxHeightRatio = 9.0 / 16.0,
+    bool useRootNavigator = false,
+    bool isDismissible = true,
+    bool enableDrag = true,
+    bool? showDragHandle,
+    bool useSafeArea = false,
+    RouteSettings? routeSettings,
+    AnimationController? transitionAnimationController,
+    Offset? anchorPoint,
+    AnimationStyle? sheetAnimationStyle,
+  }) async {
+    return showModalBottomSheet<T>(
+      context: Go.context,
+      builder: (context) => content,
+      backgroundColor: backgroundColor,
+      barrierLabel: barrierLabel,
+      elevation: elevation,
+      shape: shape,
+      clipBehavior: clipBehavior,
+      constraints: constraints,
+      barrierColor: barrierColor,
+      isScrollControlled: isScrollControlled,
+      scrollControlDisabledMaxHeightRatio: scrollControlDisabledMaxHeightRatio,
+      useRootNavigator: useRootNavigator,
+      isDismissible: isDismissible,
+      enableDrag: enableDrag,
+      showDragHandle: showDragHandle,
+      useSafeArea: useSafeArea,
+      routeSettings: routeSettings,
+      transitionAnimationController: transitionAnimationController,
+      anchorPoint: anchorPoint,
+      sheetAnimationStyle: sheetAnimationStyle,
+    );
+  }
+
+  ///If you wnat to [showSnackBar] without using context
+  ///you have to use this function and the package will pass the context automaticlly.
+  ///You can use all [showSnackBar] proparties as usuall.
+  static ScaffoldFeatureController<SnackBar, SnackBarClosedReason> snackBar(
+    SnackBar snackBar, {
+    AnimationStyle? snackBarAnimationStyle,
+  }) {
+    return ScaffoldMessenger.of(Go.context).showSnackBar(
+      snackBar,
+      snackBarAnimationStyle: snackBarAnimationStyle,
+    );
+  }
+
+  ///If you wnat to [showMaterialBanner] without using context
+  ///you have to use this function and the package will pass the context automaticlly.
+  ///You can use all [showMaterialBanner] proparties as usuall.
+  static ScaffoldFeatureController<MaterialBanner, MaterialBannerClosedReason>
+      materialBanner(MaterialBanner materialBanner) {
+    return ScaffoldMessenger.of(Go.context).showMaterialBanner(materialBanner);
+  }
+}