+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.3.5
+## 1.0.0
 
-* TODO: Describe initial release.
+* Support new Routes [DialogRoute, RawDialogRoute, PageRouteBuilder, ModalBottomSheetRoute]
+* Support passing T to Routes and return values
+* Support global access to [Anything.of(context)] using class [Gof]
+* Support show [dialog, bottomSheet, snackBar, and materialBanner] using class [GoMessenger]